Linear Regression Model
The file is explained with Simple, Polynomial and Multi Linear Regression. 
The formula for simple linear regression is
y=mx+b
where
y is the dependent variable
m is the slop
x is the independent variable
b is the intercept

It covers Mean square error and Root mean square error, overfitting and underfitting and Regularisation.